add columns without having to reset the db
To add a new column to a table, add the column to the columns array.

If the column has to be UNIQUE, you will have to reset the DB or manually add the UNIQUE constraint.

If the column has the NOT NULL constraint, you will have to set a default value, manually add the constraint and fill the values, or reset the DB.